Adding unit tests

+++ src/Microsoft.Data.SqlClient/netcore/src/Microsoft/Data/SqlClientX/Handlers/Connection/Login/FeatureExtensions.cs
@@ -13,7 +13,7 @@ namespace Microsoft.Data.SqlClientX.Handlers.Connection.Login
 {
     internal class FeatureExtensions
     {
-        public const int FeatureTerminator = 0xFF;
+        public const byte FeatureTerminator = 0xFF;
 
         public bool SessionRecoveryRequested { get; internal set; }
         public bool FederatedAuthenticationInfoRequested { get; internal set; }
